+++ flask_app/pages/dashboard.py
@@ -1,19 +1,16 @@
 from flask import render_template, request
 import sqlalchemy as sql
-import plotly.express as px
-import pandas as pd
-
 from flask_app.db import get_connection
 from flask_app.forms.experiment_chart_form import ExperimentChartForm
 from flask_app.models.study_dfs import get_experiments
+from flask_app.models.experiment_dfs import get_experiment
 from flask_app.lib.hack import get_chart_data
 
 
 def dashboard_index_page():
-    studyName = None
-    studyId = request.args.get('studyId')
-    experiments = []
-    available_growth_measurements = []
+    studyName        = None
+    studyId          = request.args.get('studyId')
+    experiment_forms = []
 
     with get_connection() as conn:
         if studyId:
@@ -46,6 +43,9 @@ def dashboard_chart_fragment():
 
         form = ExperimentChartForm(df_growth, df_reads, experiment)
 
+        # TODO (2024-09-09) Reads and FC counts are different, add another
+        # "reads" function, determine which to show based on reads CSV keys
+
         if measurement == 'Reads 16S rRNA Seq':
             figs = form.generate_reads_figures(args)
         elif measurement == 'Metabolites':
@@ -63,26 +63,6 @@ def dashboard_chart_fragment():
         ]
 
         return render_template('pages/dashboard/_figs.html', fig_htmls=fig_htmls)
-
-
-def get_experiment(experimentId, conn):
-    query = """
-        SELECT
-            E.experimentId,
-            E.experimentDescription,
-            GROUP_CONCAT(DISTINCT BRI.bioreplicateId) AS bioreplicateIds
-        FROM
-            Experiments AS E
-        LEFT JOIN
-            BioReplicatesPerExperiment AS BRI ON E.experimentUniqueId = BRI.experimentUniqueId
-        WHERE
-            E.experimentId = :experimentId
-        GROUP BY
-            E.experimentId
-    """
-    return conn.execute(sql.text(query), {'experimentId': experimentId}).one()._asdict()
-
-
 
 
 def generate_reads_figures(studyId, experimentId, args):
