+++ flask_app/templates/pages/dashboard/index.html
@@ -31,7 +31,7 @@
             <input
                 type="text"
                 name="studyId"
-                value="{{ studyId }}"
+                value="{{ studyId if studyId }}"
                 class="form-input-full form-input-blue" />
           </label>
